# docs/openapi/openapi.yaml
openapi: 3.0.0
info:
  title: MarketDataAPI
  version: '1.0'
  description: API for financial market data, instrument details, and legal entity
    information
servers:
- url: http://127.0.0.1:5000/api/v1
  description: Development server
paths:
...
  /transparency:
    get:
      summary: Get all transparency calculations with optional filtering
      description: ''
      operationId: list_transparency_calculations
      tags:
      - transparency
      parameters:
      - name: per_page
        in: query
        description: Items per page
        required: false
        schema:
          type: integer
      - name: page
        in: query
        description: Page number
        required: false
        schema:
          type: integer
      - name: isin
        in: query
        description: Filter by ISIN
        required: false
        schema:
          type: string
      - name: instrument_type
        in: query
        description: Filter by instrument type
        required: false
        schema:
          type: string
      - name: calculation_type
        in: query
        description: Filter by calculation type
        required: false
        schema:
          type: string
      - name: X-Fields
        in: header
        description: An optional fields mask
        required: false
        schema:
          type: string
          format: mask
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransparencyListResponse'
    post:
      summary: Create a new transparency calculation
      description: ''
      operationId: create_transparency_calculation
      tags:
      - transparency
      parameters:
      - name: payload
        in: body
        description: ''
        required: true
        schema:
          type: string
      responses:
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '201':
          description: Transparency calculation created
...
  /transparency/isin/{isin}:
    get:
      summary: Get transparency calculations for a specific ISIN
      description: ''
      operationId: get_transparency_by_isin
      tags:
      - transparency
      parameters:
      - name: ensure_instrument
        in: query
        description: Ensure instrument exists before creating transparency data
        required: false
        schema:
          type: boolean
      - name: calculation_type
        in: query
        description: Calculation type to get/create
        required: false
        schema:
          type: string
      - name: X-Fields
        in: header
        description: An optional fields mask
        required: false
        schema:
          type: string
          format: mask
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransparencyListResponse'
  /transparency/{transparency_id}:
    get:
      summary: Get transparency calculation by ID
      description: ''
      operationId: get_transparency_calculation
      tags:
      - transparency
      parameters:
      - name: X-Fields
        in: header
        description: An optional fields mask
        required: false
        schema:
          type: string
          format: mask
      responses:
        '404':
          description: Transparency calculation not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransparencyDetailed'
    delete:
      summary: Delete a transparency calculation
      description: ''
      operationId: delete_transparency_calculation
      tags:
      - transparency
      parameters: []
      responses:
        '404':
          description: Transparency calculation not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '200':
          description: Transparency calculation deleted
    put:
      summary: Update an existing transparency calculation
      description: ''
      operationId: update_transparency_calculation
      tags:
      - transparency
      parameters:
      - name: payload
        in: body
        description: ''
        required: true
        schema:
          type: string
      responses:
        '404':
          description: Transparency calculation not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '200':
          description: Transparency calculation updated
components:
  schemas:
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          description: Current page number
        per_page:
          type: integer
          description: Items per page
        total:
          type: integer
          description: Total number of items
...
    TransparencyListResponse:
      type: object
      properties:
        status:
          type: string
          description: Response status
          example: success
          enum:
          - success
        data:
          type: array
          items:
            $ref: '#/components/schemas/TransparencyDetailed'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
      required:
      - status
    TransparencyDetailed: {}
...
    Error:
      type: object
      properties:
        status:
          type: string
          description: Error status
          example: error
          enum:
          - error
        error:
          $ref: '#/components/schemas/ErrorDetails'
      required:
      - status
    ErrorDetails:
      type: object
      properties:
        code:
          type: string
          description: Error code
        message:
          type: string
          description: Error message description
      required:
      - code
      - message
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
security:
- ApiKeyAuth: []